backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
SQLAlchemy>=2.0.0
pydantic-settings>=2.2.1
httpx>=0.27.0
PyJWT>=2.8.0
pytest>=8.0.0
tzdata>=2024.1
